Extend Base's svdvals!. This allows e.g. cond and norm for BigFloat matrices.
Fix bug in bidiagonalization procedure for wide matrices

Suppert svdvals! for wide matrices by commenting out the check (it does
matter if the matrix is upper or lower when we only want the values)

Adjust convergence criterion. We now start with a zero shift so we should
be able to have high relative accuracy.

Remove outdated comments.